+++ Xenon/Mods/InfiniteJump.cs
@@ -1,4 +1,6 @@
-﻿using GDWeave;
+﻿// hi this is unused
+
+using GDWeave;
 using GDWeave.Godot;
 using GDWeave.Godot.Variants;
 using GDWeave.Modding;